t/xml/attlists-1.dat

#data
<!DOCTYPE a[
<!ATTLIST b>
]><a/>
#errors
2;1;empty attlist decl;;b;w
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b>
<!ATTLIST c>
]><a/>
#errors
2;1;empty attlist decl;;b;w
3;1;empty attlist decl;;c;w
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|   <!ELEMENT c >
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST bcd>
]><a/>
#errors
2;1;empty attlist decl;;bcd;w
#document
| <!DOCTYPE a>
|   <!ELEMENT bcd >
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b >
]><a/>
#errors
2;1;empty attlist decl;;b;w
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
| <{} a>

#data escaped
<!DOCTYPE a[
<!ATTLIST b\u0020
#errors
2;12;unclosed md;;;m
2;12;unclosed internal subset;;;m
2;12;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!ATTLIST bcd >
]><a/>
#errors
2;1;empty attlist decl;;bcd;w
#document
| <!DOCTYPE a>
|   <!ELEMENT bcd >
| <{} a>

#data
<!DOCTYPE a[
<!ATTLISTb>
]><a/>
#errors
2;1;empty attlist decl;;b;w
2;10;no space before md name;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
#errors
2;11;unclosed md;;;m
2;11;unclosed internal subset;;;m
2;11;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!ATTLIST >
<!ATTLIST b>
]><a/>
#errors
2;11;no md name;;;m
3;1;empty attlist decl;;b;w
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST>
<!ATTLIST b>
]><a/>
#errors
3;1;empty attlist decl;;b;w
2;10;no md name;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
| <{} a>

#data escaped
<!DOCTYPE a[
<!ATTLIST b>
<!ATTLIST\u0020
#errors
2;1;empty attlist decl;;b;w
3;10;unclosed md;;;m
3;10;unclosed internal subset;;;m
3;10;no root element;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >

#data
<!DOCTYPE a[
<!ATTLIST b>
<!ATTLIST
#errors
3;9;unclosed md;;;m
2;1;empty attlist decl;;b;w
3;9;unclosed internal subset;;;m
3;9;no root element;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >

#data
<!DOCTYPE a[
<!ATTLIS>
<!ATTLIST b>
]><a/>
#errors
2;1;bogus comment;;;m
3;1;empty attlist decl;;b;w
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
| <{} a>

#data
<!DOCTYPE a[
<!ATT c>
<!ATTLIST b>
]><a/>
#errors
2;1;bogus comment;;;m
3;1;empty attlist decl;;b;w
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
| <{} a>

#data
<!DOCTYPE a[
<!attlist c>
<!ATTLIST b>
]><a/>
#errors
2;3;lowercase keyword;ATTLIST;;m
2;1;empty attlist decl;;c;w
3;1;empty attlist decl;;b;w
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|   <!ELEMENT c >
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b>
<!ATTLIS
#errors
2;1;empty attlist decl;;b;w
3;1;bogus comment;;;m
3;8;unclosed internal subset;;;m
3;8;no root element;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >

#data
<!DOCTYPE a[
<!ATTLIST b>
<!ATTL
#errors
2;1;empty attlist decl;;b;w
3;1;bogus comment;;;m
3;6;unclosed internal subset;;;m
3;6;no root element;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >

#data
<!DOCTYPE a[
<!ATTLIST% c>
<!ATTLIST b>
]><a/>
#errors
2;1;empty attlist decl;;%;w
2;10;no space before md name;;;m
2;13;no attr type;;;m
3;1;empty attlist decl;;b;w
#document
| <!DOCTYPE a>
|   <!ELEMENT % >
|   <!ELEMENT b >
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b>
<!ATTLIST % c>
]><a/>
#errors
2;1;empty attlist decl;;b;w
3;1;empty attlist decl;;%;w
3;14;no attr type;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT % >
|   <!ELEMENT b >
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b c>
]><a/>
#errors
2;1;empty attlist decl;;b;w
2;14;no attr type;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b c
#errors
2;13;unclosed md;;;m
2;13;unclosed internal subset;;;m
2;13;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!ATTLIST b cd>
]><a/>
#errors
2;1;empty attlist decl;;b;w
2;15;no attr type;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b cd
#errors
2;14;unclosed md;;;m
2;14;unclosed internal subset;;;m
2;14;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!ATTLIST b c >
]><a/>
#errors
2;1;empty attlist decl;;b;w
2;15;no attr type;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
| <{} a>

#data escaped
<!DOCTYPE a[
<!ATTLIST b c\u0020
#errors
2;14;unclosed md;;;m
2;14;unclosed internal subset;;;m
2;14;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!ATTLIST b cd >
]><a/>
#errors
2;1;empty attlist decl;;b;w
2;16;no attr type;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
| <{} a>

#data escaped
<!DOCTYPE a[
<!ATTLIST b cd\u0020
#errors
2;15;unclosed md;;;m
2;15;unclosed internal subset;;;m
2;15;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!ATTLIST b c t>
]><a/>
#errors
2;1;empty attlist decl;;b;w
2;16;no attr default;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b c t
#errors
2;15;unclosed md;;;m
2;15;unclosed internal subset;;;m
2;15;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!ATTLIST b c tu>
]><a/>
#errors
2;1;empty attlist decl;;b;w
2;17;no attr default;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b c tu
#errors
2;16;unclosed md;;;m
2;16;unclosed internal subset;;;m
2;16;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!ATTLIST b c t >
]><a/>
#errors
2;1;empty attlist decl;;b;w
2;17;no attr default;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
| <{} a>

#data escaped
<!DOCTYPE a[
<!ATTLIST b c t\u0020
#errors
2;16;unclosed md;;;m
2;16;unclosed internal subset;;;m
2;16;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!ATTLIST b c CDATA>
]><a/>
#errors
2;1;empty attlist decl;;b;w
2;20;no attr default;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b c CDATA
#errors
2;19;unclosed md;;;m
2;19;unclosed internal subset;;;m
2;19;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!ATTLIST b c t d>
]><a/>
#errors
2;13;unknown declared type;;t;m
2;17;unquoted attr value;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c 0 () EXPLICIT "d"
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b c t d
#errors
2;17;unquoted attr value;;;m
2;17;unclosed md;;;m
2;17;unclosed internal subset;;;m
2;17;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!ATTLIST b c ID d>
]><a/>
#errors
2;18;unquoted attr value;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c ID () EXPLICIT "d"
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b c ID d
#errors
2;18;unquoted attr value;;;m
2;18;unclosed md;;;m
2;18;unclosed internal subset;;;m
2;18;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!ATTLIST b c ID "d>
]><a/>
#errors
3;3;lt in attr value;;;m
3;6;unclosed attribute value;;;m
3;6;unclosed internal subset;;;m
3;6;no root element;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c ID () EXPLICIT "d> ]><a/>"

#data
<!DOCTYPE a[
<!ATTLIST b c ID "d
#errors
2;19;unclosed attribute value;;;m
2;19;unclosed internal subset;;;m
2;19;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!ATTLIST b c ID "d">
]><a/>
#errors
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c ID () EXPLICIT "d"
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b c ID "d"
#errors
2;20;unclosed md;;;m
2;20;unclosed internal subset;;;m
2;20;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!ATTLIST b c ID "d&amp;">
]><a/>
#errors
2;24;entity not declared;;amp;;w
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c ID () EXPLICIT "d&"
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b c ID 'd>
]><a/>
#errors
3;3;lt in attr value;;;m
3;6;unclosed attribute value;;;m
3;6;unclosed internal subset;;;m
3;6;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!ATTLIST b c ID 'd
#errors
2;19;unclosed attribute value;;;m
2;19;unclosed internal subset;;;m
2;19;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!ATTLIST b c ID 'd'>
]><a/>
#errors
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c ID () EXPLICIT "d"
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b c ID 'd'
#errors
2;20;unclosed md;;;m
2;20;unclosed internal subset;;;m
2;20;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!ATTLIST b c ID 'd&amp;'>
]><a/>
#errors
2;24;entity not declared;;amp;;w
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c ID () EXPLICIT "d&"
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b c ID d&amp;>
]><a/>
#errors
2;18;unquoted attr value;;;m
2;23;entity not declared;;amp;;w
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c ID () EXPLICIT "d&"
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c CDATA "d"
e CDATA "f">
]><a/>
#errors
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c CDATA () EXPLICIT "d"
|     e CDATA () EXPLICIT "f"
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c CDATA "d"e CDATA "f">
]><a/>
#errors
3;12;no space before attr name;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c CDATA () EXPLICIT "d"
|     e CDATA () EXPLICIT "f"
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c CDATA 'd'
e CDATA "f"
>]><a/>
#errors
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c CDATA () EXPLICIT "d"
|     e CDATA () EXPLICIT "f"
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c CDATA 'd'e CDATA "f"
>]><a/>
#errors
3;12;no space before attr name;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c CDATA () EXPLICIT "d"
|     e CDATA () EXPLICIT "f"
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c CDATA d
e CDATA "f"
>]><a/>
#errors
3;9;unquoted attr value;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c CDATA () EXPLICIT "d"
|     e CDATA () EXPLICIT "f"
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c CDATA"d"
>]><a/>
#errors
3;8;no space before default value;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c CDATA () EXPLICIT "d"
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c CDATA'd'
>]><a/>
#errors
3;8;no space before default value;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c CDATA () EXPLICIT "d"
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c CDATA#REQUIRED
>]><a/>
#errors
3;8;no space before default value;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c CDATA () REQUIRED ""
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c CDATA#IMPLIED
>]><a/>
#errors
3;8;no space before default value;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c CDATA () IMPLIED ""
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c CDATA#FIXED "d"
>]><a/>
#errors
3;8;no space before default value;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c CDATA () FIXED "d"
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c CDATA #FIXED"d"
>]><a/>
#errors
3;15;no space before default value;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c CDATA () FIXED "d"
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c CDATA #IMPLIED
d CDATA #IMPLIED
>]><a/>
#errors
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c CDATA () IMPLIED ""
|     d CDATA () IMPLIED ""
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c CDATA #REQUIRED
d CDATA #IMPLIED
>]><a/>
#errors
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c CDATA () REQUIRED ""
|     d CDATA () IMPLIED ""
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c CDATA #FIXED "x"
d CDATA #IMPLIED
>]><a/>
#errors
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c CDATA () FIXED "x"
|     d CDATA () IMPLIED ""
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c CDATA #UNKNOWN
d CDATA #IMPLIED
>]><a/>
#errors
3;1;unknown default type;;UNKNOWN;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c CDATA () 0 ""
|     d CDATA () IMPLIED ""
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c IDREF #IMPLIED
>]><a/>
#errors
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c IDREF () IMPLIED ""
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c IDREFS #IMPLIED
>]><a/>
#errors
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c IDREFS () IMPLIED ""
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c NMTOKEN #IMPLIED
>]><a/>
#errors
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c NMTOKEN () IMPLIED ""
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c NMTOKENS #IMPLIED
>]><a/>
#errors
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c NMTOKENS () IMPLIED ""
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c ENTITY #IMPLIED
>]><a/>
#errors
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c ENTITY () IMPLIED ""
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c ENTITIES #IMPLIED
>]><a/>
#errors
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c ENTITIES () IMPLIED ""
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c NOTATION #IMPLIED
>]><a/>
#errors
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c NOTATION () IMPLIED ""
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c ENUMERATION #IMPLIED
>]><a/>
#errors
3;1;unknown declared type;;ENUMERATION;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c 0 () IMPLIED ""
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c cdata #IMPLIED
>]><a/>
#errors
3;1;unknown declared type;;cdata;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c 0 () IMPLIED ""
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c CDATA #FIXED 'd'
e CDATA #IMPLIED
>
]><a/>
#errors
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c CDATA () FIXED "d"
|     e CDATA () IMPLIED ""
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c CDATA #FIXED'd'
e CDATA #IMPLIED
>
]><a/>
#errors
3;15;no space before default value;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c CDATA () FIXED "d"
|     e CDATA () IMPLIED ""
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c CDATA #FIXED 'd'e CDATA #IMPLIED
>
]><a/>
#errors
3;19;no space before attr name;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c CDATA () FIXED "d"
|     e CDATA () IMPLIED ""
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c CDATA #x 'd'
e CDATA #IMPLIED
>
]><a/>
#errors
3;1;unknown default type;;x;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c CDATA () 0 "d"
|     e CDATA () IMPLIED ""
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c CDATA # 'd'
e CDATA #IMPLIED
>
]><a/>
#errors
2;1;empty attlist decl;;b;w
3;10;no default type;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c () #IMPLIED
>
]><a/>
#errors
3;4;empty allowed token;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c ENUMERATION () IMPLIED ""
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c NOTATION () #IMPLIED
>
]><a/>
#errors
3;13;empty allowed token;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c NOTATION () IMPLIED ""
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c ()#IMPLIED
>
]><a/>
#errors
3;4;empty allowed token;;;m
3;5;no space before default value;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c ENUMERATION () IMPLIED ""
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c NOTATION ()#IMPLIED
>
]><a/>
#errors
3;13;empty allowed token;;;m
3;14;no space before default value;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c NOTATION () IMPLIED ""
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c() #IMPLIED
>
]><a/>
#errors
3;2;no space before paren;;;m
3;3;empty allowed token;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c ENUMERATION () IMPLIED ""
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c NOTATION() #IMPLIED
>
]><a/>
#errors
3;11;no space before paren;;;m
3;12;empty allowed token;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c NOTATION () IMPLIED ""
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c () "x"
>
]><a/>
#errors
3;4;empty allowed token;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c ENUMERATION () EXPLICIT "x"
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c NOTATION () "x"
>
]><a/>
#errors
3;13;empty allowed token;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c NOTATION () EXPLICIT "x"
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c ()"x"
>
]><a/>
#errors
3;4;empty allowed token;;;m
3;5;no space before default value;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c ENUMERATION () EXPLICIT "x"
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c NOTATION ()"x"
>
]><a/>
#errors
3;13;empty allowed token;;;m
3;14;no space before default value;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c NOTATION () EXPLICIT "x"
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c (y) "x"
>
]><a/>
#errors
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c ENUMERATION (y) EXPLICIT "x"
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c (yz) "x"
>
]><a/>
#errors
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c ENUMERATION (yz) EXPLICIT "x"
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c (y|z) "x"
>
]><a/>
#errors
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c ENUMERATION (y|z) EXPLICIT "x"
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c (y | z | abc) "x"
>
]><a/>
#errors
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c ENUMERATION (y|z|abc) EXPLICIT "x"
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c (   ABC  |DEF|G  ) "x"
>
]><a/>
#errors
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c ENUMERATION (ABC|DEF|G) EXPLICIT "x"
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c (|) "x"
>
]><a/>
#errors
3;4;empty allowed token;;;m
3;5;empty allowed token;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c ENUMERATION () EXPLICIT "x"
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c (|y) "x"
>
]><a/>
#errors
3;4;empty allowed token;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c ENUMERATION (y) EXPLICIT "x"
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c (y|) "x"
>
]><a/>
#errors
3;6;empty allowed token;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c ENUMERATION (y) EXPLICIT "x"
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c (y||z) "x"
>
]><a/>
#errors
3;6;empty allowed token;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c ENUMERATION (y|z) EXPLICIT "x"
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c (||) "x"
>
]><a/>
#errors
3;4;empty allowed token;;;m
3;5;empty allowed token;;;m
3;6;empty allowed token;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c ENUMERATION () EXPLICIT "x"
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c (
#errors
3;3;unclosed md;;;m
3;3;unclosed internal subset;;;m
3;3;no root element;;;m
#document
| <!DOCTYPE a>

#data escaped
<!DOCTYPE a[
<!ATTLIST b
c (\u0020
#errors
3;4;unclosed md;;;m
3;4;unclosed internal subset;;;m
3;4;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!ATTLIST b
c (x
#errors
3;4;unclosed md;;;m
3;4;unclosed internal subset;;;m
3;4;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!ATTLIST b
c (xy
#errors
3;5;unclosed md;;;m
3;5;unclosed internal subset;;;m
3;5;no root element;;;m
#document
| <!DOCTYPE a>

#data escaped
<!DOCTYPE a[
<!ATTLIST b
c (x\u0020
#errors
3;5;unclosed md;;;m
3;5;unclosed internal subset;;;m
3;5;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!ATTLIST b
c (x|
#errors
3;5;unclosed md;;;m
3;5;unclosed internal subset;;;m
3;5;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!ATTLIST b
c (>
]><a/>
#errors
3;4;unclosed allowed tokens;;;m
#document
| <!DOCTYPE a>
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c ( >
]><a/>
#errors
3;5;unclosed allowed tokens;;;m
#document
| <!DOCTYPE a>
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c (a>
]><a/>
#errors
3;5;unclosed allowed tokens;;;m
#document
| <!DOCTYPE a>
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c (ab>
]><a/>
#errors
3;6;unclosed allowed tokens;;;m
#document
| <!DOCTYPE a>
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c (x >
]><a/>
#errors
3;6;unclosed allowed tokens;;;m
#document
| <!DOCTYPE a>
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c (x)
#errors
3;5;no root element;;;m
3;5;unclosed internal subset;;;m
3;5;unclosed md;;;m
#document
| <!DOCTYPE a>

#data escaped
<!DOCTYPE a[
<!ATTLIST b
c (x)\u0009
#errors
3;6;no root element;;;m
3;6;unclosed internal subset;;;m
3;6;unclosed md;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!ATTLIST b
c (x|>
]><a/>
#errors
3;6;unclosed allowed tokens;;;m
#document
| <!DOCTYPE a>
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c (x y) #IMPLIED>
]><a/>
#errors
3;5;space in allowed token;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c ENUMERATION (x y) IMPLIED ""
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c (x y  |z) #IMPLIED>
]><a/>
#errors
3;5;space in allowed token;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c ENUMERATION (x y|z) IMPLIED ""
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c (x y   z) #IMPLIED>
]><a/>
#errors
3;5;space in allowed token;;;m
3;9;space in allowed token;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c ENUMERATION (x y z) IMPLIED ""
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c (x     y) #IMPLIED>
]><a/>
#errors
3;9;space in allowed token;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c ENUMERATION (x y) IMPLIED ""
| <{} a>

#data
<!DOCTYPE a[
<!attlist b>
]><a/>
#errors
2;1;empty attlist decl;;b;w
2;3;lowercase keyword;ATTLIST;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
| <{} a>

#data
<!DOCTYPE a[
<!ATTLISt b>
]><a/>
#errors
2;1;empty attlist decl;;b;w
2;3;lowercase keyword;ATTLIST;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
| <{} a>

#data
<!DOCTYPE a[
<!aTTLIST b>
]><a/>
#errors
2;1;empty attlist decl;;b;w
2;3;lowercase keyword;ATTLIST;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b>
<!ATTLIST b>
]><a/>
#errors
2;1;empty attlist decl;;b;w
3;1;duplicate attlist decl;;b;w
3;1;empty attlist decl;;b;w
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
| <{} a>

#data
<!DOCTYPE a[
<!ELEMENT b ANY>
<!ATTLIST b>
]><a/>
#errors
3;1;empty attlist decl;;b;w
#document
| <!DOCTYPE a>
|   <!ELEMENT b ANY>
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b>
<!ELEMENT b ANY>
]><a/>
#errors
2;1;empty attlist decl;;b;w
#document
| <!DOCTYPE a>
|   <!ELEMENT b ANY>
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c CDATA #IMPLIED
c CDATA #REQUIRED>
]><a/>
#errors
4;1;duplicate attrdef;;c;w
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c CDATA () IMPLIED ""
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b
c CDATA #IMPLIED>
<!ATTLIST b
c CDATA #REQUIRED>
]><a/>
#errors
4;1;duplicate attlist decl;;b;w
5;1;duplicate attrdef;;c;w
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c CDATA () IMPLIED ""
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b c CDATA #IMPLIED>
%x;
<!ATTLIST d e CDATA #IMPLIED>
]><a/>
#errors
3;1;stop processing;;;i
3;2;string in internal subset;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c CDATA () IMPLIED ""
| <{} a>

#data
<?xml version="1.0" standalone="yes"?>
<!DOCTYPE a[
<!ATTLIST b c CDATA #IMPLIED>
%x;
<!ATTLIST d e CDATA #IMPLIED>
]><a/>
#errors
4;2;string in internal subset;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c CDATA () IMPLIED ""
|   <!ELEMENT d >
|     e CDATA () IMPLIED ""
| <{} a>

#data escaped
<!DOCTYPE a[
<!ATTLIST b c CDATA "ab\u0020\u0009\u000C\u000D\u000Acd">
]><a/>
#errors
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c CDATA () EXPLICIT "ab    cd"
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b c CDATA "ab&#xd;&#xa;&#x9;&#x20;&#xc;">
]>
<a/>
#errors
#document escaped
| <!DOCTYPE a>
|   <!ELEMENT b >
|     c CDATA () EXPLICIT "ab\u000d\u000a\u0009\u0020\u000c"
| <{} a>
